updated entities, added controllers for user actions

+++ deech.me.api/controllers/BookmarkController.cs
@@ -1,25 +1,33 @@
-using AutoMapper;
+using System.Linq;
+
+using Microsoft.AspNetCore.Mvc;
+
 using deech.me.data.entities;
 using deech.me.logic.abstractions;
 using deech.me.logic.models;
-using Microsoft.AspNetCore.Mvc;
+
+using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 
 namespace deech.me.api.controllers
 {
+    [Authorize]
     [Route("[controller]")]
     public class BookmarkController : ControllerBase
     {
         private readonly IMapper _mapper;
         private readonly IWriteDataService<Bookmark> _bookmarkDataService;
+
         public BookmarkController(IWriteDataService<Bookmark> bookmarkDataService, IMapper mapper)
         {
-            _bookmarkDataService = bookmarkDataService;
-            _mapper = mapper;
+            this._bookmarkDataService = bookmarkDataService;
+            this._mapper = mapper;
         }
 
         [HttpPost]
         public IActionResult Add([FromBody] BookmarkModel bookmark)
         {
+            bookmark.UserId = GetUserId();
             var entity = _mapper.Map<Bookmark>(bookmark);
             entity = _bookmarkDataService.Add(entity);
             var result = _mapper.Map<BookmarkModel>(entity);
@@ -30,26 +38,42 @@ namespace deech.me.api.controllers
         [HttpPut]
         public IActionResult Update([FromBody] BookmarkModel bookmark)
         {
+            if (bookmark.UserId != GetUserId())
+            {
+                return BadRequest();
+            }
+
             var entity = _mapper.Map<Bookmark>(bookmark);
             entity = _bookmarkDataService.Update(entity);
             var result = _mapper.Map<BookmarkModel>(entity);
+
             return Ok(result);
         }
 
         [HttpDelete]
         public IActionResult Delete([FromBody] BookmarkModel bookmark)
         {
+            if (bookmark.UserId != GetUserId())
+            {
+                return BadRequest();
+            }
+
             var entity = _mapper.Map<Bookmark>(bookmark);
             _bookmarkDataService.Delete(entity);
+
             return Ok();
         }
 
         [HttpGet]
         public IActionResult Get(int bookId)
         {
-            var bookmarks = _bookmarkDataService.GetMultiple(b => b.Paragraph.BookId == bookId, 0, 10000);
+            var userId = GetUserId();
+            var bookmarks = _bookmarkDataService.GetMultiple(b => b.Paragraph.BookId == bookId && b.UserId == userId, 0, 10000);
             var result = _mapper.Map<BookmarkModel>(bookmarks);
+
             return Ok(result);
         }
+
+        private string GetUserId() => HttpContext.User.Claims.ToList().Single(c => c.Type == "id").Value;
     }
 }
