+++ deech.me.api/controllers/CitationController.cs
@@ -1,25 +1,33 @@
-using AutoMapper;
+using System.Linq;
+
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
 using deech.me.data.entities;
 using deech.me.logic.abstractions;
 using deech.me.logic.models;
-using Microsoft.AspNetCore.Mvc;
+
+using AutoMapper;
 
 namespace deech.me.api.controllers
 {
+    [Authorize]
     [Route("[controller]")]
     public class CitationController : ControllerBase
     {
         private readonly IMapper _mapper;
         private readonly IWriteDataService<Citation> _citationDataService;
+
         public CitationController(IWriteDataService<Citation> citationDataService, IMapper mapper)
         {
-            _citationDataService = citationDataService;
-            _mapper = mapper;
+            this._citationDataService = citationDataService;
+            this._mapper = mapper;
         }
 
         [HttpPost]
         public IActionResult Add([FromBody] CitationModel citation)
         {
+            citation.UserId = GetUserId();
             var entity = _mapper.Map<Citation>(citation);
             entity = _citationDataService.Add(entity);
             var result = _mapper.Map<CitationModel>(entity);
@@ -30,26 +38,42 @@ namespace deech.me.api.controllers
         [HttpPut]
         public IActionResult Update([FromBody] CitationModel citation)
         {
+            if (citation.UserId != GetUserId())
+            {
+                return BadRequest();
+            }
+
             var entity = _mapper.Map<Citation>(citation);
             entity = _citationDataService.Update(entity);
             var result = _mapper.Map<CitationModel>(entity);
+
             return Ok(result);
         }
 
         [HttpDelete]
         public IActionResult Delete([FromBody] CitationModel citation)
         {
+            if (citation.UserId != GetUserId())
+            {
+                return BadRequest();
+            }
+
             var entity = _mapper.Map<Citation>(citation);
             _citationDataService.Delete(entity);
+
             return Ok();
         }
 
         [HttpGet]
         public IActionResult Get(int bookId)
         {
-            var citations = _citationDataService.GetMultiple(b => b.Paragraph.BookId == bookId, 0, 10000);
+            var userId = GetUserId();
+            var citations = _citationDataService.GetMultiple(b => b.Paragraph.BookId == bookId && b.UserId == userId, 0, 10000);
             var result = _mapper.Map<CitationModel>(citations);
+
             return Ok(result);
         }
+
+        private string GetUserId() => HttpContext.User.Claims.ToList().Single(c => c.Type == "id").Value;
     }
 }
